move MetricsConfig to CytosisConfig
